+++ src/components/forms/index.js
@@ -1,4 +1,18 @@
 
+// import { createUser as createUserAPI } from "../../api/users";
 
 
 
+// export const onSubmit = async (data) => {
+//     setIsSubmitting(true);
+//     // setError(null);
+//     try {
+//       const user = await createUserAPI(data);
+//       setList((prevList) => [...prevList, user]);
+//       reset(); // Clear form inputs
+//     } catch (error) {
+//     //   setError(error.message || "An unexpected error occurred. Please try again.");
+//     } finally {
+//       setIsSubmitting(false);
+//     }
+//   };
